C++ InspectorPackagerConnection: Indicate new implementation in `/json` response (#42200)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42200

Changelog: [Internal]

Quick hack to make it easy to determine whether a given build of React Native is using the C++ implementation of InspectorPackagerConnection or the legacy platform-specific implementation.

For now, we just append this information to the `title` field. Ultimately, rather than polluting the title, this should be an inert capability flag that gets reported via `inspector-proxy`. I'm not doing that yet since we have work in the pipeline to set up a proper capability flag system soon.

Reviewed By: huntie

Differential Revision: D52629415

fbshipit-source-id: a4e873f4be78ae49b35b94fd5d41d0e2efc02dbe